Removes solana interfaces
