+++ base58/build.gradle.kts
@@ -36,7 +36,6 @@ kotlin {
     sourceSets {
         val commonMain by getting {
             dependencies {
-                //put your multiplatform dependencies here
                 implementation(libs.crypto)
             }
         }
