+++ rpc/src/commonMain/kotlin/foundation/metaplex/rpc/Cluster.kt
@@ -1,4 +1,4 @@
-package foundation.metaplex.solanainterfaces
+package foundation.metaplex.rpc
 
 
 /**
