+++ settings.gradle.kts
@@ -19,8 +19,7 @@ include(":solana")
 include(":solanapublickeys")
 include(":base58")
 include(":solanaeddsa")
-include(":solanainterfaces")
-include(":solanakeypair")
 include(":amount")
 include(":readapi")
 include(":rpc")
+include(":signer")
