+++ signer/build.gradle.kts
@@ -4,7 +4,6 @@ plugins {
     alias(libs.plugins.kotlin.multiplatform)
     alias(libs.plugins.android.library)
     alias(libs.plugins.maven.publish)
-    kotlin("plugin.serialization") version "1.9.0"
 }
 
 @OptIn(org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi::class)
@@ -29,7 +28,7 @@ kotlin {
         macosArm64()
     ).forEach {
         it.binaries.framework {
-            baseName = "SolanaInterfaces"
+            baseName = "RPC"
             xcf.add(this)
         }
     }
@@ -38,22 +37,18 @@ kotlin {
         val commonMain by getting {
             dependencies {
                 implementation(project(mapOf("path" to ":solanapublickeys")))
-                implementation(project(mapOf("path" to ":amount")))
             }
         }
         val commonTest by getting {
             dependencies {
                 implementation(kotlin("test"))
-                implementation(libs.kotlinx.coroutines.test)
             }
         }
-        val jvmMain by getting
-        val jvmTest by getting
     }
 }
 
 android {
-    namespace = "foundation.metaplex.solanainterfaces"
+    namespace = "com.metaplex.signer"
     compileSdk = 33
     defaultConfig {
         minSdk = 24
@@ -65,5 +60,5 @@ android {
 }
 
 mavenPublishing {
-    coordinates(group as String, "solanainterfaces", version as String)
+    coordinates(group as String, "signer", version as String)
 }
