+++ solana/src/commonMain/kotlin/foundation/metaplex/solana/transactions/MessageInterface.kt
@@ -1,4 +1,4 @@
-package foundation.metaplex.solanainterfaces
+package foundation.metaplex.solana.transactions
 
 import foundation.metaplex.solanapublickeys.PublicKey
