--- solanakeypair/build.gradle.kts
@@ -1,68 +1,0 @@
-import org.jetbrains.kotlin.gradle.plugin.mpp.apple.XCFramework
-
-plugins {
-    alias(libs.plugins.kotlin.multiplatform)
-    alias(libs.plugins.android.library)
-    alias(libs.plugins.maven.publish)
-}
-
-@OptIn(org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi::class)
-kotlin {
-    targetHierarchy.default()
-
-    androidTarget {
-        compilations.all {
-            kotlinOptions {
-                jvmTarget = "17"
-            }
-        }
-    }
-    jvm()
-
-    val xcf = XCFramework()
-    listOf(
-        iosX64(),
-        iosArm64(),
-        iosSimulatorArm64(),
-        macosX64(),
-        macosArm64()
-    ).forEach {
-        it.binaries.framework {
-            baseName = "SolanaKeypair"
-            xcf.add(this)
-        }
-    }
-
-    sourceSets {
-        val commonMain by getting {
-            dependencies {
-                //put your multiplatform dependencies here
-                implementation(project(mapOf("path" to ":solanainterfaces")))
-                implementation(project(mapOf("path" to ":solanapublickeys")))
-            }
-        }
-        val commonTest by getting {
-            dependencies {
-                implementation(kotlin("test"))
-            }
-        }
-        val jvmMain by getting
-        val jvmTest by getting
-    }
-}
-
-android {
-    namespace = "foundation.metaplex.solanakeypair"
-    compileSdk = 33
-    defaultConfig {
-        minSdk = 24
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_17
-        targetCompatibility = JavaVersion.VERSION_17
-    }
-}
-
-mavenPublishing {
-    coordinates(group as String, "solanakeypair", version as String)
-}
