--- solanakeypair/src/commonMain/kotlin/foundation/metaplex/solanakeypair/SolanaKeypair.kt
@@ -1,8 +1,0 @@
-package foundation.metaplex.solanakeypair
-import foundation.metaplex.solanainterfaces.Keypair
-import foundation.metaplex.solanapublickeys.PublicKey
-
-class SolanaKeypair(
-    override val publicKey: PublicKey,
-    override val secretKey: ByteArray
-) : Keypair
